Updated album and music interfaces to const folder. Updated album to include quantity of songs. Added Album to swagger and also fixed music text in swagger as it was about football.
