+++ modules/music/consts/Album.ts
@@ -1,0 +1,10 @@
+interface Album {
+    id: string
+    name: string
+    releasedate: string
+    price: string
+    publisher: string
+    songs: number
+}
+
+export type { Album };
