+++ modules/music/consts/Music.ts
@@ -1,0 +1,7 @@
+interface Music {
+    id: string
+    genre: string
+    song: string
+}
+
+export type { Music };
